iOS: Design up the remote control UI, hook up actions.
There is currently a problem with more than one action getting fired, I blame the xcode integration.

+++ Core/Model/Player.cs
@@ -26,7 +26,7 @@ namespace GmusicbrowserRemote.Core
         /// </value>
         public float? Volume { get; set; }
 
-        public double? PlayPosition { get; set; }
+        public float? PlayPosition { get; set; }
 
         public Player () {
         }
